+++ packages/observable-client/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Improve logic of `withOptionalHash` for non-static blocks.
+- Add types of topics for SystemEvent
 
 ## 0.5.10 - 2024-10-16